services/code-analysis/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
psycopg2-binary==2.9.9
redis==5.0.1
boto3==1.34.0
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
ast-decompiler==0.7.0
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.4
tree-sitter-java==0.20.5
tree-sitter-go==0.20.0
gitpython==3.1.40
docker==6.1.3
aiofiles==23.2.1
asyncpg==0.29.0
celery==5.3.4